+++ Basis/Packages/com.basis.server/BasisNetworkCore/BasisPacketUtil.cs
@@ -6,7 +6,7 @@ namespace BasisNetworkCore
 {
     public static class BasisPacketUtil
     {
-        public static bool ValidatePacket(byte previous, byte ValidateMe)
+        public static bool ValidatePacket(ushort previous, ushort ValidateMe)
         {
             if (IsNewer(previous, ValidateMe) && previous != ValidateMe)
             {
@@ -15,7 +15,7 @@ namespace BasisNetworkCore
             return false;
         }
         // Returns true if seq1 is newer than seq2
-        public static bool IsNewer(byte seq1, byte seq2)
+        public static bool IsNewer(ushort seq1, ushort seq2)
         {
             return (byte)(seq1 - seq2) < 128;
         }
